STYLE Updated fonts
Updated fonts to planed design for current coming soon landing
page. Recreated logo to close replacement to recreation. Used the
the logo font in headlines, Tenso. Then used VT323 from Google
Fonts. There is no bold version of the VT323 font and when making
it bold by the browser the m letter is too condensed horizontally.
As a fix I had to set a class to make the bold font wider. An
issue with this fix it makes all characters wider in bold text. We
only need the m's to be wider. Later I will need to come up with a
fix that only selects m characters. Also found a old reset from
before adding normalize.css. This override normalize.css and
needed to be removed and was removed.