BUILD Add Blog Pages
Added static blog pages. With a new UI component for tags. Also
required a couple new TS libraries and a npm package. The
removeHTML library is a dependency of the getReadingTime library
to remove all html tags from content. The getReadingTime library
takes the content and gets the average time to read the content.
It uses the newly added npm package "reading-time" to accomplish
this. In the process I made some TypeScript fixes and added
aliases. I also updated all import references to use the aliases.